requirements.txt
Flask
Flask-CORS
google-generativeai
gunicorn
# ใส่ library อื่นๆ ที่คุณใช้